ha_assist/requirements.txt
requests
jsonschema
json-repair
homeassistant-api
python-dotenv
